LibGfx/JPEGXL: Move some functions to a common header
The decoder, while being far from complete is already a huge file. Let's
split it in multiple files. This commit is the first step toward this
goal.

No functional change.
